Output more information in MSP_BOARD_INFO (#5206)
* Append the target name to the response of MSP_BOARD_INFO

This allows MSP clients to retrieve the actual target name, which
is useful for e.g. automatically upgrading the firmware on a board
without asking the user to select the target manually or to present
a warning dialog if we detect the user has chosen the wrong target.

* Output wether a board uses VCP in MSP_BOARD_INFO

This allows the configurator to detect if the board uses VCP
without using an external list of boards.

* Use a single bit to output wether the board uses VCP

This lets us use the remaining 7 bits for other flags.
Also, output if the board has softserial support using the next
bit.