Fix dip in D weight at center when transition is 0
Currently transition cannot be disabled.  Nearest thing is to set it to
0.01 because of a divide by zero issue in the code.
0.01 doesn’t disable transition, it actually cuts D weight sharply to
zero when sticks are centred - a sharp cut when passing the stick
across +/- 1% of stick movement.
This PR allows transition to be set to 0, totally disabling it.
Reduces CPU usage and makes weight consistent across all stick
movements.