Capture and use crash flip if set initially when arming is delayed due to beacon Previously if arming was delayed due to DSHOT beacon then the arming type (normal or crash-flip) would be captured *after* the delay. This change captures the state on entering the delay which more accurately represents the pilot's intent.
For example, the pilot may have crash-flip on a momentary switch that's initially set when the arming switch is enabled. Previously if the pilot didn't hold the momentary switch through the entire delay the quad would arm in normal mode. This change uses the state when the arming switch first enables so even if the momentary crash-flip switch is released the quad will still arm in crash-flip mode.