Added back the `OK` status, tried to keep a similar spirit of the code changes
The `OK` status is it shows the craft is ready to arm (no errors and not waiting for GPS).  The flight modes are only when armed, showing the currently active flight mode.  This way, it's clear when armed and when not armed.  This confirmation of armed/unarmed is important to trigger events and modes.  For example, when to start a timer or monitor signal strength.