ff_boost
Adds to FF a high-passed  FF element.  Sort of like 'jerk' or 'acceleration' or second derivative of stick input.
Benefit is earlier rise in FF, followed by a drop-off.  A bit like throttle boost, this leads to quicker rise in FF at the start of a move, with pull-back to zero FF sooner than otherwise, and negative FF when the sticks stop moving entirely.  This improves tracking accuracy during the onset of quick unputs, and at the same time reduces overshoot.