Fix default axis values for arbitrary gyro alignment Prevents clean installs from displaying the sensor alignment as changed values in the `diff`.
Sets the default values for the sensor alignment axes to match the selected sensor alignment. Previously the axis values defaulted to 0 so when the active value was set based on the sensor alignment these values would appear as changes in the `diff` even though they were actually the default values. Corrected the default values to be based on the selected sensor alignment default.