EXST - Use the .bin file as the source for the hex file.
* This bypasses the problem introduced when the .data section was moved
from RAM to DTCM_RAM.
* This also ensures that the hash in the .hex file will be valid when
the hex is flashed to a NAND chip as the hash is generated from the .bin
file content.  If the hex contains gaps between sections and the gaps
are erased on the NAND flash then erased sectors are full of 0xFF, not
0x00, which will mean the computed hash won't match.

EXST - Add note regarding currently supported bootloader requirements
regarding the CODE_RAM address.

EXST - Fix make 'dirty' file issue.

When `make` is run twice, the TARGET_BIN is recreated incorrectly.

Root cause was that objcopy actually modifies the INPUT file when no
output file is specified, even when the operation is to extract a
section from the input file...

EXST - Improve messaging when generating EXST bin files.

This should help with understanding of the exact operations required for
EXST bin file generation.

Added warning about hard coded start address.