Set default dyanmic notch range to medium
In 4.0x the dynamic notch range was set by default to AUTO mode.

This would set range to low, medium or high based on the dynamic lowpass max setting, which the user was supposed to configure to suit the maximum rpm of the quad.

In 4.1 we are moving the maximum value of the dynamic gyro lowpass very high whenever the rpm filter is enabled.  And, generally, it is OK to do this even on non-rpm quads.

The problem then becomes that the dynamic notch quietly shifts itself to HIGH mode, and that can cause it to not track ordinary frame and prop resonances like it should.

I think it will be better overall in medium mode, which works best for most quads out of the box.  Users of very high RPM quads can choose high.  XClass quads or machines with low frequency resonance should choose low, and should set the dyn_notch_min_hz value low enough to reach the resonant peak.