Add cliDebugPrint functions to facilitate easy debug printing to CLI To use, include `cli/cli_debug_print.h` in your code and be sure `USE_CLI_DEBUG_PRINT` is defined. Then you'll have access to the following functions to print debugging messages in the CLI: ``` cliDebugPrintLineFeed cliDebugPrint cliDebugPrintLine cliDebugPrintf cliDebugPrintLinef ``` Output is suppressed when the CLI is not open.
May interfere with the autocomplete initialization when first entering the CLI if your code is outputting data when the CLI first opens. But as this is only meant for debugging it shouldn't be much of a concern.

You may also need to rate limit your messages if printing data in a loop.

All of the debugging code must be removed from any completed code before submission.